+++ tests/rule_tests/mod.rs
@@ -132,6 +132,7 @@ macro_rules! test_rule_wihout_effects {
 
 mod compute_expression;
 mod convert_index_to_field;
+mod filter_early_return;
 mod group_local_assignment;
 mod inject_value;
 mod no_local_function;